move things to the right dir
