+++ service/src/main/java/bio/terra/workspace/service/admin/flights/cloudcontexts/gcp/SyncGcpIamRolesFlight.java
@@ -1,4 +1,4 @@
-package bio.terra.workspace.service.workspace.flight.cloudcontext.gcp;
+package bio.terra.workspace.service.admin.flights.cloudcontexts.gcp;
 
 import static bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.GCP_PROJECT_IDS;
 import static java.util.Objects.requireNonNull;
