+++ service/src/test/java/bio/terra/workspace/service/admin/AdminServiceTest.java
@@ -13,12 +13,12 @@ import bio.terra.workspace.common.BaseConnectedTest;
 import bio.terra.workspace.connected.UserAccessUtils;
 import bio.terra.workspace.connected.WorkspaceConnectedTestUtils;
 import bio.terra.workspace.db.WorkspaceDao;
+import bio.terra.workspace.service.admin.flights.cloudcontexts.gcp.GcpIamCustomRolePatchStep;
+import bio.terra.workspace.service.admin.flights.cloudcontexts.gcp.RetrieveGcpIamCustomRoleStep;
 import bio.terra.workspace.service.crl.CrlService;
 import bio.terra.workspace.service.iam.model.WsmIamRole;
 import bio.terra.workspace.service.job.JobService;
 import bio.terra.workspace.service.resource.controlled.cloud.gcp.CustomGcpIamRole;
-import bio.terra.workspace.service.workspace.flight.cloudcontext.gcp.GcpIamCustomRolePatchStep;
-import bio.terra.workspace.service.workspace.flight.cloudcontext.gcp.RetrieveGcpIamCustomRoleStep;
 import bio.terra.workspace.service.workspace.model.CloudPlatform;
 import bio.terra.workspace.service.workspace.model.GcpCloudContext;
 import com.google.api.services.iam.v1.model.Role;
